feat: added delete

+++ ExtendibleHashing.cpp
@@ -68,7 +68,7 @@ void ExtendibleHashing<T>::splitBucket(int depth, string key) {
     int bucketIndex = directory[key];
     auto *bucket = loadBucket(bucketIndex);
 
-    if (bucket->localDepth >= globalDepth) {
+    if (bucket->localDepth >= globalDepth - 1) {
         addOverflowBucket(bucketIndex);
         return;
     }
@@ -116,6 +116,9 @@ void ExtendibleHashing<T>::splitBucket(int depth, string key) {
 template<typename T>
 void ExtendibleHashing<T>::addOverflowBucket(int index) {
     auto bucket = loadBucket(index);
+    if (bucket->size < 1) {
+        return;
+    }
     auto newBucket = new Bucket(bucket->localDepth);
 
     newBucket->insert(bucket->removeLast());
@@ -125,6 +128,129 @@ void ExtendibleHashing<T>::addOverflowBucket(int index) {
     saveBucket(bucket, index);
     saveBucket(newBucket, buckets++);
 }
+
+template<class T>
+void ExtendibleHashing<T>::deleteItem(T record) {
+    bitset<sizeof(int) * 8> hashValue = hash(record);
+    bitset<sizeof(int) * 8> currentSearch = hashValue & bitset<32>(1);
+
+    int currentDepth = 0;
+
+    while( (directory.find(currentSearch.to_string().substr((sizeof(int) * 8) - (currentDepth + 1), currentDepth + 1 )) == directory.end()) && (currentDepth < globalDepth - 1) ) {
+        currentDepth++;
+        currentSearch = (hashValue & bitset<32>((1 << (currentDepth + 1)) - 1));
+    }
+
+    int index = directory[currentSearch.to_string().substr((sizeof(int) * 8) - (currentDepth + 1), currentDepth + 1)];
+
+    auto bucket = loadBucket(index);
+
+    bool deleted = false;
+    if (bucket->contains(record)) {
+        bucket->remove(record);
+        saveBucket(bucket, index);
+        deleted = true;
+    }
+    int previousIndex = index;
+    if (!deleted) {
+        if (bucket->next != -1) {
+            while (bucket->next != -1 && !bucket->contains(record)) {
+                previousIndex = index;
+                index = bucket->next;
+                bucket = loadBucket(bucket->next);
+            }
+        }
+        if (!bucket->contains(record)) {
+            return;
+        }
+        bucket->remove(record);
+        saveBucket(bucket, index);
+    }
+    if (bucket->isEmpty()) {
+        cout << "Bucket is empty" << endl;
+        if (previousIndex != index) {
+            auto previousBucket = loadBucket(previousIndex);
+            previousBucket->next = bucket->next;
+            saveBucket(previousBucket, previousIndex);
+        }
+        else {
+            bool done = false;
+            // Check if sibling bucket (same key up to its most significant bit (ie 101 and 001 are siblings)) has records
+            while (!done && currentDepth > 0) {
+                string key = currentSearch.to_string().substr((sizeof(int) * 8) - (currentDepth + 1), currentDepth + 1);
+                string siblingKey = key;
+                siblingKey[0] = (siblingKey[0] == '0') ? '1' : '0';
+                int siblingIndex = directory[siblingKey];
+                auto siblingBucket = loadBucket(siblingIndex);
+
+                // if both buckets are empty, we can delete them and make a parent key (one with less depth (ie. 01 is a parent for 001 and 101)) an entry in the directory
+                // we can reuse the lowest index of one of the siblings (to avoid creating a new bucket in memory)
+                int lowestIndex = (index < siblingIndex) ? index : siblingIndex;
+                if (siblingBucket->isEmpty()) {
+                    directory.erase(key);
+                    directory.erase(siblingKey);
+                    directory[key.substr(1, key.size() - 1)] = lowestIndex;
+                    delete siblingBucket;
+                    delete bucket;
+                    currentDepth--;
+                } else {
+                    done = true;
+                }
+            }
+        }
+    }
+    if (bucket->next != -1) {
+        cout << "Bucket has next" << endl;
+        // we can bring an item from its next bucket to the current bucket
+        auto nextBucket = loadBucket(bucket->next);
+        bucket->insert(nextBucket->removeLast());
+        if (nextBucket->isEmpty()) {
+            bucket->next = nextBucket->next;
+            saveBucket(bucket, index);
+            delete nextBucket;
+        } else {
+            saveBucket(bucket, index);
+            saveBucket(nextBucket, bucket->next);
+        }
+    }
+    if (currentDepth != 0 && previousIndex == index && bucket->next == -1) {
+        cout << "Possible merge" << endl;
+        // We might want to merge the bucket with a sibling into its parent.
+        bool done = false;
+        while (!done && currentDepth > 0) {
+            string key = currentSearch.to_string().substr((sizeof(int) * 8) - (currentDepth + 1), currentDepth + 1);
+            string siblingKey = key;
+            siblingKey[0] = (siblingKey[0] == '0') ? '1' : '0';
+            int siblingIndex = directory[siblingKey];
+            auto siblingBucket = loadBucket(siblingIndex);
+
+            if (siblingBucket->size + bucket->size <= maxBucketSize) {
+                // Parent bucket is not a bucket as it's been split before
+                // We can reuse the lowest index of the sibling and merge the two buckets
+                int lowestIndex = (index < siblingIndex) ? index : siblingIndex;
+                auto parentBucket = new Bucket();
+                parentBucket->localDepth = currentDepth - 1;
+                parentBucket->size = siblingBucket->size + bucket->size;
+                for (int i = 0; i < siblingBucket->size; i++) {
+                    parentBucket->insert(siblingBucket->records[i]);
+                }
+                for (int i = 0; i < bucket->size; i++) {
+                    parentBucket->insert(bucket->records[i]);
+                }
+                saveBucket(parentBucket, lowestIndex);
+                directory.erase(key);
+                directory.erase(siblingKey);
+                directory[key.substr(1, key.size() - 1)] = lowestIndex;
+                delete siblingBucket;
+                delete bucket;
+                currentDepth--;
+            } else {
+                done = true;
+            }
+        }
+    }
+}
+
 
 bitset<sizeof(int) * 8> CustomHash(int record) {
         return bitset<sizeof(int) * 8>(record % 3);
@@ -150,4 +276,10 @@ int main(){
 
         cout << endl;
     }
+    cout << "Deleting 6" << endl;
+    eh.deleteItem(6);
+    cout << "Buckets in directory:" << endl;
+    eh.printAllBucketsFromDir();
+    cout << "Buckets in memory:" << endl;
+    eh.printAllBucketsFromMemory();
 }
